Fix adding a new todo
